+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Sentry disabled in development environment
 * Data loading functions refactored into utils module and refactored to return consistent data structures
 * external data files aligned to use snake case slugs
+* Refactored citation section to use a record instance as a single prop
 
 ## [0.5.3] - 2023-12-30